Unify the serialport.list output
- [osx] remove the 0x from productID and locationId and vendorId
- [linux] remove the starting 0x from any values, small refactor
- [linux] Fixup output using the hex encoded fields
- [windows] Grab the serial number from the pnp id
- Isolates list to it's own platform files (c++)
- Better ensure our serialports fields are present
