+++ .docs/README.hbs
@@ -405,15 +405,15 @@ $ serialport-list -h
 
 
 $ serialport-list
-/dev/cu.Bluetooth-Incoming-Port
-/dev/cu.usbmodem1421    Arduino (www.arduino.cc)
+/dev/tty.Bluetooth-Incoming-Port
+/dev/tty.usbmodem1421    Arduino (www.arduino.cc)
 
 $ serialport-list -f json
-[{"comName":"/dev/cu.Bluetooth-Incoming-Port"},{"comName":"/dev/cu.usbmodem1421","manufacturer":"Arduino (www.arduino.cc)","serialNumber":"752303138333518011C1","locationId":"0x14200000","vendorId":"0x2341","productId":"0x0043"}]
+[{"comName":"/dev/tty.Bluetooth-Incoming-Port"},{"comName":"/dev/tty.usbmodem1421","manufacturer":"Arduino (www.arduino.cc)","serialNumber":"752303138333518011C1","locationId":"14200000","vendorId":"2341","productId":"0043"}]
 
 $ serialport-list -f jsonline
-{"comName":"/dev/cu.Bluetooth-Incoming-Port"}
-{"comName":"/dev/cu.usbmodem1421","manufacturer":"Arduino (www.arduino.cc)","serialNumber":"752303138333518011C1","locationId":"0x14200000","vendorId":"0x2341","productId":"0x0043"}
+{"comName":"/dev/tty.Bluetooth-Incoming-Port"}
+{"comName":"/dev/tty.usbmodem1421","manufacturer":"Arduino (www.arduino.cc)","serialNumber":"752303138333518011C1","locationId":"14200000","vendorId":"2341","productId":"0043"}
 ```
 
 ### Serial Port Terminal
@@ -439,8 +439,8 @@ $ serialport-term -h
     --echo --localecho             Print characters as you type them
 
 $ serialport-term -l
-/dev/cu.Bluetooth-Incoming-Port
-/dev/cu.usbmodem1421    Arduino (www.arduino.cc)
+/dev/tty.Bluetooth-Incoming-Port
+/dev/tty.usbmodem1421    Arduino (www.arduino.cc)
 ```
 
 ### Serial Port Repl
@@ -448,31 +448,31 @@ $ serialport-term -l
 
 You can make use of the `serialport-repl` command with;
 ```bash
-serialport-repl # to auto detect an arduino
-serialport-repl /path/name # to connect to a specific port
+$ serialport-repl # to auto detect an arduino
+$ serialport-repl /dev/tty.usbmodem1421 # to connect to a specific port
 ```
 
 It will load a serialport object with debugging turned on.
 ```
   serialport:binding:auto-detect loading DarwinBinding +0ms
-port = SerialPort("/path/name", { autoOpen: false })
+port = SerialPort("/dev/tty.usbmodem1421", { autoOpen: false })
 globals { SerialPort, portName, port }
 > SerialPort.list()
   serialport:main .list +6s
-[ { comName: '/path/name',
-    manufacturer: undefined,
-    serialNumber: undefined,
+[ { comName: '/dev/tty.usbmodem1421',
+    manufacturer: 'Arduino (www.arduino.cc)',
+    serialNumber: '752303138333518011C1',
     pnpId: undefined,
-    locationId: undefined,
-    vendorId: undefined,
-    productId: undefined } ]
+    locationId: '14200000',
+    vendorId: '2341',
+    productId: '0043' } ]
 > port.write('Calling all Autobots!')
 true
 > port.read()
   serialport:main _read queueing _read for after open +1m
 null
 > port.open()
-  serialport:main opening path: serialport-repl +30s
+  serialport:main opening path: /dev/tty.usbmodem1421 +30s
   serialport:bindings open +1ms
 ```
