+++ lib/bindings/linux-list.js
@@ -1,60 +1,97 @@
 'use strict';
 
 const childProcess = require('child_process');
-const LineStream = require('../parsers/readline');
+const Readline = require('../parsers/readline');
 
-function checkPathAndDevice(path) {
-  // get only serial port names
+// get only serial port names
+function checkPathOfDevice(path) {
   return (/(tty(S|ACM|USB|AMA|MFD)|rfcomm)/).test(path) && path;
 }
+
 function propName(name) {
   return {
     'DEVNAME': 'comName',
-    'ID_VENDOR': 'manufacturer',
-    'ID_SERIAL': 'serialNumber',
+    'ID_VENDOR_ENC': 'manufacturer',
+    'ID_SERIAL_SHORT': 'serialNumber',
     'ID_VENDOR_ID': 'vendorId',
-    'ID_MODEL_ID': 'productId',
+    'ID_MODEL_ENC': 'productId',
     'DEVLINKS': 'pnpId'
   }[name.toUpperCase()];
 }
+
+function decodeHexEscape(str) {
+  return str.replace(/\\x([a-fA-F0-9]{2})/g, (a, b) => {
+    return String.fromCharCode(parseInt(b, 16));
+  });
+}
+
 function propVal(name, val) {
-  if (/productId|vendorId/.test(name) && !/^0x/.test(val)) {
-    return `0x${val}`;
-  }
   if (name === 'pnpId') {
-    return val.split(' ')
-      .map((path) => path.match(/\/by-id\/(.*)/))
-      .filter(a => a).map(a => a[1])[0];
+    const match = val.match(/\/by-id\/([^\s]+)/);
+    return (match && match[1]) || undefined;
+  }
+  if (name === 'manufacturer') {
+    return decodeHexEscape(val);
+  }
+  if (name === 'productId') {
+    return decodeHexEscape(val);
+  }
+  if (/^0x/.test(val)) {
+    return val.substr(2);
   }
   return val;
 }
 
 function listLinux() {
   return new Promise((resolve, reject) => {
-    const ude = childProcess.spawn('udevadm', ['info', '-e']);
-    ude.on('error', reject);
-    const lines = new LineStream();
     const ports = [];
-    let obj, prop, name;
-    lines.on('finish', () => resolve(ports));
+    const ude = childProcess.spawn('udevadm', ['info', '-e']);
+    const lines = ude.stdout.pipe(new Readline());
+    ude.on('error', reject);
     lines.on('error', reject);
-    lines.on('data', (data) => {
-      const line = data.toString();
-      if ((name = line.match(/^N: (.*)/))) {
-        if (checkPathAndDevice(name[1])) {
-          ports.push(obj = {});
+
+    let port = {};
+    let skipPort = false;
+    lines.on('data', (line) => {
+      const lineType = line.slice(0, 1);
+      const data = line.slice(3);
+      // new port entry
+      if (lineType === 'P') {
+        port = {
+          manufacturer: undefined,
+          serialNumber: undefined,
+          pnpId: undefined,
+          locationId: undefined,
+          vendorId: undefined,
+          productId: undefined
+        };
+        skipPort = false;
+        return;
+      }
+
+      if (skipPort) { return }
+
+      // Check dev name and save port if it matches flag to skip the rest of the data if not
+      if (lineType === 'N') {
+        if (checkPathOfDevice(data)) {
+          ports.push(port);
+        } else {
+          skipPort = true;
         }
-      } else
-      if ((name = line.match(/^E: (.*)=(.*)/))) {
-        if (obj && (prop = propName(name[1]))) {
-          obj[prop] = propVal(prop, name[2]);
-        }
-      } else
-      if (/^P: /.test(line)) {
-        obj = undefined;
+        return;
+      }
+
+      // parse data about each port
+      if (lineType === 'E') {
+        const keyValue = data.match(/^(.+)=(.*)/);
+        if (!keyValue) { return }
+        const key = propName(keyValue[1]);
+        if (!key) { return }
+        port[key] = propVal(key, keyValue[2]);
       }
     });
-    ude.stdout.pipe(lines);
+
+    lines.on('finish', () => resolve(ports));
   });
 }
