+++ lib/bindings/win32.js
@@ -5,7 +5,17 @@ const promisify = require('../util').promisify;
 
 class WindowsBinding extends BaseBinding {
   static list() {
-    return promisify(binding.list)();
+    return promisify(binding.list)().then(ports => {
+      // Grab the serial number from the pnp id
+      ports.forEach(port => {
+        if (port.pnpId) {
+          const parts = port.pnpId.match(/USB\\(.+)\\(.+)/);
+          if (!parts) { return }
+          port.serialNumber = parts.pop();
+        }
+      });
+      return ports;
+    });
   }
 
   constructor(opt) {
