+++ lib/serialport.js
@@ -553,22 +553,45 @@ SerialPort.prototype.drain = function(callback) {
 
 /**
  * Retrieves a list of available serial ports with metadata. Only the `comName` is guaranteed. If unavailable the other fields will be undefined. The `comName` is either the path or an identifier (eg `COM1`) used to open the SerialPort.
+ *
+ * We make an effort to identify the hardware attached and have consistent results between systems. Linux and OS X are mostly consistent. Windows relies on 3rd party device drivers for the information and is unable to guarantee the information. On windows If you have a USB connected device can we provide a serial number otherwise it will be `undefined`. The `pnpId` and `locationId` are not the same or present on all systems. The examples below were run with the same Arduino Uno.
  * @type {function}
  * @param {listCallback=} callback
  * @returns {Promise} Resolves with the list of available serial ports.
  * @example
 ```js
-// example port information
+// OSX example port
 {
-  comName: '/dev/cu.usbmodem1421',
+  comName: '/dev/tty.usbmodem1421',
   manufacturer: 'Arduino (www.arduino.cc)',
-  serialNumber: '757533138333964011C1',
+  serialNumber: '752303138333518011C1',
   pnpId: undefined,
-  locationId: '0x14200000',
-  vendorId: '0x2341',
-  productId: '0x0043'
+  locationId: '14500000',
+  productId: '0043',
+  vendorId: '2341'
 }
 
+// Linux example port
+{
+  comName: '/dev/ttyACM0',
+  manufacturer: 'Arduino (www.arduino.cc)',
+  serialNumber: '752303138333518011C1',
+  pnpId: 'usb-Arduino__www.arduino.cc__0043_752303138333518011C1-if00',
+  locationId: undefined,
+  productId: '0043',
+  vendorId: '2341'
+}
+
+// Windows example port
+{
+  comName: 'COM3',
+  manufacturer: 'Arduino LLC (www.arduino.cc)',
+  serialNumber: '752303138333518011C1',
+  pnpId: 'USB\\VID_2341&PID_0043\\752303138333518011C1',
+  locationId: 'Port_#0003.Hub_#0001',
+  productId: '0043',
+  vendorId: '2341'
+}
 ```
 
 ```js
@@ -581,6 +604,7 @@ SerialPort.list(function (err, ports) {
     console.log(port.manufacturer);
   });
 });
+
 // promise approach
 SerialPort.list()
   .then(ports) {...});
