+++ src/serialport_unix.cpp
@@ -1,4 +1,6 @@
+#include "./serialport_unix.h"
 #include "./serialport.h"
+
 #include <sys/file.h>
 #include <unistd.h>
 #include <fcntl.h>
@@ -8,13 +10,6 @@
 #ifdef __APPLE__
 #include <AvailabilityMacros.h>
 #include <sys/param.h>
-#include <IOKit/IOKitLib.h>
-#include <IOKit/IOCFPlugIn.h>
-#include <IOKit/usb/IOUSBLib.h>
-#include <IOKit/serial/IOSerialKeys.h>
-
-uv_mutex_t list_mutex;
-Boolean lockInitialised = FALSE;
 #endif
 
 #if defined(MAC_OS_X_VERSION_10_4) && (MAC_OS_X_VERSION_MIN_REQUIRED >= MAC_OS_X_VERSION_10_4)
@@ -31,8 +26,6 @@ Boolean lockInitialised = FALSE;
 #include <linux/serial.h>
 #endif
 
-int ToBaudConstant(int baudRate);
-int ToDataBitsConstant(int dataBits);
 int ToStopBitsConstant(SerialPortStopBits stopBits);
 
 int ToBaudConstant(int baudRate) {
@@ -73,23 +66,6 @@ int ToBaudConstant(int baudRate) {
   }
   return -1;
 }
-
-#ifdef __APPLE__
-typedef struct SerialDevice {
-    char port[MAXPATHLEN];
-    char locationId[MAXPATHLEN];
-    char vendorId[MAXPATHLEN];
-    char productId[MAXPATHLEN];
-    char manufacturer[MAXPATHLEN];
-    char serialNumber[MAXPATHLEN];
-} stSerialDevice;
-
-typedef struct DeviceListItem {
-    struct SerialDevice value;
-    struct DeviceListItem *next;
-    int* length;
-} stDeviceListItem;
-#endif
 
 int ToDataBitsConstant(int dataBits) {
   switch (dataBits) {
@@ -332,296 +308,6 @@ void EIO_Close(uv_work_t* req) {
   if (-1 == close(data->fd)) {
     snprintf(data->errorString, sizeof(data->errorString), "Error: %s, unable to close fd %d", strerror(errno), data->fd);
   }
-}
-
-#ifdef __APPLE__
-
-// Function prototypes
-static kern_return_t FindModems(io_iterator_t *matchingServices);
-static io_service_t GetUsbDevice(io_service_t service);
-static stDeviceListItem* GetSerialDevices();
-
-
-static kern_return_t FindModems(io_iterator_t *matchingServices) {
-    kern_return_t     kernResult;
-    CFMutableDictionaryRef  classesToMatch;
-    classesToMatch = IOServiceMatching(kIOSerialBSDServiceValue);
-    if (classesToMatch != NULL) {
-        CFDictionarySetValue(classesToMatch,
-                             CFSTR(kIOSerialBSDTypeKey),
-                             CFSTR(kIOSerialBSDAllTypes));
-    }
-
-    kernResult = IOServiceGetMatchingServices(kIOMasterPortDefault, classesToMatch, matchingServices);
-
-    return kernResult;
-}
-
-static io_service_t GetUsbDevice(io_service_t service) {
-  IOReturn status;
-  io_iterator_t   iterator = 0;
-  io_service_t    device = 0;
-
-  if (!service) {
-    return device;
-  }
-
-  status = IORegistryEntryCreateIterator(service,
-                                         kIOServicePlane,
-                                         (kIORegistryIterateParents | kIORegistryIterateRecursively),
-                                         &iterator);
-
-  if (status == kIOReturnSuccess) {
-    io_service_t currentService;
-    while ((currentService = IOIteratorNext(iterator)) && device == 0) {
-      io_name_t serviceName;
-      status = IORegistryEntryGetNameInPlane(currentService, kIOServicePlane, serviceName);
-      if (status == kIOReturnSuccess && IOObjectConformsTo(currentService, kIOUSBDeviceClassName)) {
-        device = currentService;
-      } else {
-        // Release the service object which is no longer needed
-        (void) IOObjectRelease(currentService);
-      }
-    }
-
-    // Release the iterator
-    (void) IOObjectRelease(iterator);
-  }
-
-  return device;
-}
-
-static void ExtractUsbInformation(stSerialDevice *serialDevice, IOUSBDeviceInterface  **deviceInterface) {
-  kern_return_t kernResult;
-  UInt32 locationID;
-  kernResult = (*deviceInterface)->GetLocationID(deviceInterface, &locationID);
-  if (KERN_SUCCESS == kernResult) {
-    snprintf(serialDevice->locationId, 11, "0x%08x", locationID);
-  }
-
-  UInt16 vendorID;
-  kernResult = (*deviceInterface)->GetDeviceVendor(deviceInterface, &vendorID);
-  if (KERN_SUCCESS == kernResult) {
-    snprintf(serialDevice->vendorId, 7, "0x%04x", vendorID);
-  }
-
-  UInt16 productID;
-  kernResult = (*deviceInterface)->GetDeviceProduct(deviceInterface, &productID);
-  if (KERN_SUCCESS == kernResult) {
-    snprintf(serialDevice->productId, 7, "0x%04x", productID);
-  }
-}
-
-static stDeviceListItem* GetSerialDevices() {
-  kern_return_t kernResult;
-  io_iterator_t serialPortIterator;
-  char bsdPath[MAXPATHLEN];
-
-  FindModems(&serialPortIterator);
-
-  io_service_t modemService;
-  kernResult = KERN_FAILURE;
-  Boolean modemFound = false;
-
-  // Initialize the returned path
-  *bsdPath = '\0';
-
-  stDeviceListItem* devices = NULL;
-  stDeviceListItem* lastDevice = NULL;
-  int length = 0;
-
-  while ((modemService = IOIteratorNext(serialPortIterator))) {
-    CFTypeRef bsdPathAsCFString;
-    bsdPathAsCFString = IORegistryEntrySearchCFProperty(
-      modemService,
-      kIOServicePlane,
-      CFSTR(kIODialinDeviceKey),
-      kCFAllocatorDefault,
-      kIORegistryIterateRecursively);
-
-    if (bsdPathAsCFString) {
-      Boolean result;
-
-      // Convert the path from a CFString to a C (NUL-terminated)
-      result = CFStringGetCString((CFStringRef) bsdPathAsCFString,
-                    bsdPath,
-                    sizeof(bsdPath),
-                    kCFStringEncodingUTF8);
-      CFRelease(bsdPathAsCFString);
-
-      if (result) {
-        stDeviceListItem *deviceListItem = (stDeviceListItem*) malloc(sizeof(stDeviceListItem));
-        stSerialDevice *serialDevice = &(deviceListItem->value);
-        strcpy(serialDevice->port, bsdPath);
-        memset(serialDevice->locationId, 0, sizeof(serialDevice->locationId));
-        memset(serialDevice->vendorId, 0, sizeof(serialDevice->vendorId));
-        memset(serialDevice->productId, 0, sizeof(serialDevice->productId));
-        serialDevice->manufacturer[0] = '\0';
-        serialDevice->serialNumber[0] = '\0';
-        deviceListItem->next = NULL;
-        deviceListItem->length = &length;
-
-        if (devices == NULL) {
-          devices = deviceListItem;
-        } else {
-          lastDevice->next = deviceListItem;
-        }
-
-        lastDevice = deviceListItem;
-        length++;
-
-        modemFound = true;
-        kernResult = KERN_SUCCESS;
-
-        uv_mutex_lock(&list_mutex);
-
-        io_service_t device = GetUsbDevice(modemService);
-
-        if (device) {
-          CFStringRef manufacturerAsCFString = (CFStringRef) IORegistryEntryCreateCFProperty(device,
-                      CFSTR(kUSBVendorString),
-                      kCFAllocatorDefault,
-                      0);
-
-          if (manufacturerAsCFString) {
-            Boolean result;
-            char    manufacturer[MAXPATHLEN];
-
-            // Convert from a CFString to a C (NUL-terminated)
-            result = CFStringGetCString(manufacturerAsCFString,
-                          manufacturer,
-                          sizeof(manufacturer),
-                          kCFStringEncodingUTF8);
-
-            if (result) {
-              strcpy(serialDevice->manufacturer, manufacturer);
-            }
-
-            CFRelease(manufacturerAsCFString);
-          }
-
-          CFStringRef serialNumberAsCFString = (CFStringRef) IORegistryEntrySearchCFProperty(device,
-                      kIOServicePlane,
-                      CFSTR(kUSBSerialNumberString),
-                      kCFAllocatorDefault,
-                      kIORegistryIterateRecursively);
-
-          if (serialNumberAsCFString) {
-            Boolean result;
-            char    serialNumber[MAXPATHLEN];
-
-            // Convert from a CFString to a C (NUL-terminated)
-            result = CFStringGetCString(serialNumberAsCFString,
-                          serialNumber,
-                          sizeof(serialNumber),
-                          kCFStringEncodingUTF8);
-
-            if (result) {
-              strcpy(serialDevice->serialNumber, serialNumber);
-            }
-
-            CFRelease(serialNumberAsCFString);
-          }
-
-          IOCFPlugInInterface **plugInInterface = NULL;
-          SInt32        score;
-          HRESULT       res;
-
-          IOUSBDeviceInterface  **deviceInterface = NULL;
-
-          kernResult = IOCreatePlugInInterfaceForService(device, kIOUSBDeviceUserClientTypeID, kIOCFPlugInInterfaceID,
-                               &plugInInterface, &score);
-
-          if ((kIOReturnSuccess != kernResult) || !plugInInterface) {
-            continue;
-          }
-
-          // Use the plugin interface to retrieve the device interface.
-          res = (*plugInInterface)->QueryInterface(plugInInterface, CFUUIDGetUUIDBytes(kIOUSBDeviceInterfaceID),
-                               (LPVOID*) &deviceInterface);
-
-          // Now done with the plugin interface.
-          (*plugInInterface)->Release(plugInInterface);
-
-          if (res || deviceInterface == NULL) {
-            continue;
-          }
-
-          // Extract the desired Information
-          ExtractUsbInformation(serialDevice, deviceInterface);
-
-          // Release the Interface
-          (*deviceInterface)->Release(deviceInterface);
-
-          // Release the device
-          (void) IOObjectRelease(device);
-        }
-
-        uv_mutex_unlock(&list_mutex);
-      }
-    }
-
-    // Release the io_service_t now that we are done with it.
-    (void) IOObjectRelease(modemService);
-  }
-
-  IOObjectRelease(serialPortIterator);  // Release the iterator.
-
-  return devices;
-}
-
-#endif
-
-void EIO_List(uv_work_t* req) {
-  ListBaton* data = static_cast<ListBaton*>(req->data);
-
-#ifndef __APPLE__
-  // This code exists in javascript for other unix platforms
-  snprintf(data->errorString, sizeof(data->errorString), "List is not Implemented");
-  return;
-# else
-  if (!lockInitialised) {
-    uv_mutex_init(&list_mutex);
-    lockInitialised = TRUE;
-  }
-
-  stDeviceListItem* devices = GetSerialDevices();
-  if (*(devices->length) > 0) {
-    stDeviceListItem* next = devices;
-
-    for (int i = 0, len = *(devices->length); i < len; i++) {
-      stSerialDevice device = (* next).value;
-
-      ListResultItem* resultItem = new ListResultItem();
-      resultItem->comName = device.port;
-
-      if (*device.locationId) {
-        resultItem->locationId = device.locationId;
-      }
-      if (*device.vendorId) {
-        resultItem->vendorId = device.vendorId;
-      }
-      if (*device.productId) {
-        resultItem->productId = device.productId;
-      }
-      if (*device.manufacturer) {
-        resultItem->manufacturer = device.manufacturer;
-      }
-      if (*device.serialNumber) {
-        resultItem->serialNumber = device.serialNumber;
-      }
-      data->results.push_back(resultItem);
-
-      stDeviceListItem* current = next;
-
-      if (next->next != NULL) {
-        next = next->next;
-      }
-
-      free(current);
-    }
-  }
-#endif
 }
 
 void EIO_Set(uv_work_t* req) {
