+++ src/serialport_win.h
@@ -1,8 +1,8 @@
 #ifndef SRC_SERIALPORT_WIN_H_
 #define SRC_SERIALPORT_WIN_H_
-#include <stdio.h>
-#include <stdlib.h>
-#include <string.h>
+// #include <stdio.h>
+// #include <stdlib.h>
+// #include <string.h>
 #include <nan.h>
 #include <list>
 #include <string>
@@ -39,4 +39,25 @@ struct ReadBaton {
 NAN_METHOD(Read);
 void EIO_Read(uv_work_t* req);
 void EIO_AfterRead(uv_work_t* req);
+
+NAN_METHOD(List);
+void EIO_List(uv_work_t* req);
+void EIO_AfterList(uv_work_t* req);
+
+struct ListResultItem {
+  std::string comName;
+  std::string manufacturer;
+  std::string serialNumber;
+  std::string pnpId;
+  std::string locationId;
+  std::string vendorId;
+  std::string productId;
+};
+
+struct ListBaton {
+  Nan::Callback callback;
+  std::list<ListResultItem*> results;
+  char errorString[ERROR_STRING_SIZE];
+};
+
 #endif  // SRC_SERIALPORT_WIN_H_
