+++ test/initializers/assert.js
@@ -1,1 +1,4 @@
-global.assert = require('chai').assert;
+const chai = require('chai');
+const chaiSubset = require('chai-subset');
+global.assert = chai.assert;
+chai.use(chaiSubset);
